refactor(mirror): mark reverse builder registry + autoload discipline (audits #5, #17)
Two related cleanups:

#5 — Mark reverse direction now mirrors the node-side OCP pattern.
MirrorToCoreModel#apply_mark used to switch on a SIMPLE_MARKS hash
plus a case statement for link/xref. New MarkReverseBuilder module
holds one Builder class per mark type, self-registering via the same
`registers` DSL as ReverseBuilder. Adding a new mark type is now
purely additive, with mark-level parity to node-level dispatch.

#17 — Consolidate reverse_builder.rb + reverse_builders.rb into one
file. The plural file only existed to host a `require_relative`
bridge to the singular module file. Per project autoload discipline
(no require_relative for internal library code), the autoload target
must be self-contained: the consolidated reverse_builder.rb holds
the module, the Base class, and every built-in Builder subclass.
Same shape for the new mark_reverse_builder.rb. mirror_to_core_model.rb
no longer needs any require_relative — referencing ReverseBuilder
and MarkReverseBuilder triggers autoload from mirror.rb.

Specs:
- reverse_builder_spec: replace `allow(node).to receive(:type)` with
  a real Struct (audit #1 — no doubles).
- mark_reverse_builder_spec: new — REGISTRY coverage, lookup, build
  dispatch, Base contract, integration, OCP extension.
